requirements.txt
# AURA - Autonomous User-side Robotic Assistant
# Python Dependencies

# Core Dependencies
requests>=2.31.0          # HTTP requests for API communication
python-dotenv>=1.0.0      # Environment variable management

# Computer Vision and Screen Capture
mss>=9.0.1               # Multi-Screen Screenshot for screen capture
Pillow>=10.0.0           # Image processing and manipulation
opencv-python>=4.8.0     # Computer vision operations (optional)

# GUI Automation
PyAutoGUI>=0.9.54        # Desktop automation and control
pygetwindow>=0.0.9       # Window management utilities

# Audio Processing
openai-whisper>=20231117  # Speech-to-text recognition
pvporcupine>=3.0.0       # Wake word detection (Picovoice)
pygame>=2.5.2            # Audio playback for sound effects
pydub>=0.25.1            # Audio file manipulation
sounddevice>=0.4.6       # Audio input/output
numpy>=1.24.0            # Numerical operations for audio

# Text-to-Speech Options
pyttsx3>=2.90            # Cross-platform TTS (system TTS)
# piper-tts>=1.2.0       # Alternative: Piper TTS (uncomment if preferred)

# Machine Learning and AI
torch>=2.0.0             # PyTorch for local model inference
transformers>=4.35.0     # Hugging Face transformers
accelerate>=0.24.0       # Model acceleration utilities

# Web and API
fastapi>=0.104.0         # Web framework (for future web interface)
uvicorn>=0.24.0          # ASGI server (for future web interface)
websockets>=12.0         # WebSocket support (for future real-time features)

# Data Processing
pandas>=2.1.0            # Data manipulation and analysis
jsonschema>=4.19.0       # JSON schema validation

# Logging and Monitoring
structlog>=23.2.0        # Structured logging
psutil>=5.9.0            # System and process utilities

# Development and Testing
pytest>=7.4.0           # Testing framework
pytest-asyncio>=0.21.0  # Async testing support
pytest-mock>=3.12.0     # Mocking utilities
black>=23.9.0            # Code formatting
flake8>=6.1.0            # Code linting
mypy>=1.6.0              # Type checking

# Optional: Advanced Features
# scikit-learn>=1.3.0    # Machine learning utilities
# matplotlib>=3.8.0     # Plotting and visualization
# seaborn>=0.13.0        # Statistical visualization

# Platform-specific dependencies
# Windows
pywin32>=306; sys_platform == "win32"     # Windows API access
pywinutils>=0.1.0; sys_platform == "win32" # Windows utilities

# macOS
pyobjc>=10.0; sys_platform == "darwin"    # macOS API access
AppKit>=0.2.8; sys_platform == "darwin"   # macOS application framework

# Linux
python-xlib>=0.33; sys_platform == "linux" # X11 utilities for Linux
pycairo>=1.25.0; sys_platform == "linux"   # Cairo graphics library